+++ libs/spark/src/Unstable_ListItem/Unstable_ListItem.stories.tsx
@@ -264,8 +264,6 @@ export const EnrollmentDetail = (args) => {
 const useMicroschoolThumbnailListItemStyles = makeStyles((theme) => ({
   root: {
     paddingBlockStart: 24,
-  },
-  rootWithSecondaryAction: {
     '@media (max-width: 320px)': {
       paddingInlineEnd: 16,
     },