+++ libs/spark/CHANGELOG.md
@@ -28,6 +28,8 @@
   - Removed props-dynamic styles.
 - **Unstable_Link**
   - Removed props-dynamic styles.
+- **Unstable_ListItem**
+  - Removed props-dynamic styles.
 - **Unstable_Menu**
   - Props API Changes:
     - `PaperProps`: see **Unstable_Paper**.
@@ -105,6 +107,17 @@
     - `rootDarker` -> ` ` (remove) (move styles to `root`)
     - `rootInset` -> ` ` (remove) (move styles to `root`)
     - `rootMiddle` -> ` ` (remove) (move styles to `root`)
+- **Unstable_Divider**
+  - CSS API:
+    - `rootWithSecondaryAction`: removed
+    - `button`: removed
+    - `nest`: removed
+    - `secondaryActionFlexStart`: removed
+  - Migration:
+    - `rootWithSecondaryAction` -> ` ` (remove) (move styles to `root`)
+    - `button` -> ` ` (remove) (move styles to `root`)
+    - `nest` -> ` ` (remove) (move styles to `root`)
+    - `secondaryActionFlexStart` -> ` ` (remove) (move styles to `secondaryAction`)
 - **Unstable_Paper**
   - Props API:
     - `elevation`: removed values 1..24.